team12-22-main/backend/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
contourpy==1.0.7
crispy-bootstrap5==0.7
cycler==0.11.0
Django==4.1.7
django-bootstrap-v5==1.0.11
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-mysql==2.2.0
djangorestframework==3.14.0
fonttools==4.39.0
gunicorn==20.1.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
threadpoolctl==3.1.0
